% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_statistical_par.R
\name{calculate_RSD}
\alias{calculate_RSD}
\title{Title}
\usage{
calculate_RSD(exp, pdata)
}
\arguments{
\item{exp}{data.frame: row as sample, colum as metabolite(or features/variables)}

\item{pdata}{data.frame: row-samples, column-phenotype variables or factors.}
}
\value{
data.frame includes RSD of every variable within pre-defined groups.
}
\description{
Title
}
