% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consensus_Cluster_Analysis.R
\name{Consensus_Cluster_Analysis}
\alias{Consensus_Cluster_Analysis}
\title{Consensus_Cluster_Analysis}
\usage{
Consensus_Cluster_Analysis(
  Exp,
  k,
  methods = "ALL",
  dist.method = "euclidean",
  methods.consensus = "average",
  scale = TRUE,
  plot.cluster.pca = FALSE,
  saveplot = FALSE,
  plot.dir.suffix = NULL,
  ConsensusClusterPlus.reps = 100
)
}
\arguments{
\item{Exp}{data.frame or matrix with rownames represent samples, colnames represent features like proteins or genes.}

\item{k}{number of clusters}

\item{methods}{Charactor vector. Default "ALL" will perform all 20 clustering methods.}

\item{dist.method}{Method for calculating dissimilarity. Default "euclidean", can be any methods supported by dist()}

\item{methods.consensus}{Method used for the final consensus cluster by hieraichical clustering. Default are set to "average".}

\item{scale}{Logical value, whether to scale your input data. Default: TRUE.}

\item{plot.cluster.pca}{Logical value, whether to plot pca of every clustering results. Default: FALSE.}

\item{saveplot}{under developing}

\item{plot.dir.suffix}{under developing}

\item{ConsensusClusterPlus.reps}{Integer value, used by consensusClusterPlus.}
}
\value{
List contain all cluster results. Also return consensus cluster heatmap and PCA plot indicating the consensus cluster result.
}
\description{
Consensus clustering result of multi-clustering algorithms.
}
\examples{
consensus_clusters <- Consensus_Cluster_Analysis(iris[-5], 3, methods = "ALL")
plot_consensus_clusters(consensus_clusters)

consensus_clusters <- Consensus_Cluster_Analysis(iris[-5], 3, methods = c("kmeans", "pam"))

}
