% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/DEA_Consensus_Differential_Expression_Analysis.R
\name{Consensus_Differential_Expression}
\alias{Consensus_Differential_Expression}
\alias{DEA_ANOVA}
\title{Consensus_Differential_Expression}
\usage{
Consensus_Differential_Expression(Exp, groups, adj.method = "BH")
}
\arguments{
\item{Exp}{data.frame or matrix with rownames represent samples, colnames represent features like proteins or genes.}

\item{groups}{Charactor or factor vector indicates group information. In two-way anova, groups are a data.frame, each column indicate one group factor.}

\item{type}{"one-way" or "two-way" anova.}
}
\value{
\code{Consensus_Differential_Expression}
One list including all results of Differential Expression analysis
\code{DEA_ANOVA}
Anova result with pairwised comparation.
}
\description{
Consensus_Differential_Expression
}
\examples{
res.anova <- DEA_ANOVA(iris[-5], groups = iris$Species)
}
