% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_regressionAnalysis.R
\name{DATA_lmregression}
\alias{DATA_lmregression}
\title{Regression analysis}
\usage{
DATA_lmregression(
  pdata,
  exp,
  pr.variable,
  variable.cofounder = NULL,
  resp.variable
)
}
\arguments{
\item{resp.variable}{}
}
\value{

}
\description{
Regression analysis
}
