% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_imputation.R
\name{fill_NA}
\alias{fill_NA}
\title{Title}
\usage{
fill_NA(exp, pdata, method = "knn", constant = 100, na = NA)
}
\arguments{
\item{method}{min0.5 (half of the minimum positive value), min0.2, knn, constant(0,), bpca, svd,}

\item{na}{}
}
\value{

}
\description{
Title
}
