% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_normalization_visualization.R
\name{normalize_data_visualization}
\alias{normalize_data_visualization}
\title{Title}
\usage{
normalize_data_visualization(
  exp,
  pdata,
  plot,
  order.fixed = TRUE,
  type = "default",
  variable.name = "variable",
  expression.unit = "expression",
  traj.single.metabo.RSD.quantile = 0.1,
  log = TRUE
)
}
\arguments{
\item{exp}{data.frame: row as sample, colum as metabolite(or features/variables)}

\item{pdata}{data.frame: row as sample, colum as phenotype data(or factors)}

\item{plot}{type of plot, including: boxplot, RSD, PCA, traj}

\item{type}{"default": display all samples or "onlyQC": only display QC samples in boxplot.}

\item{traj.single.metabo.RSD.quantile}{RSD quantile for drawing the RSD boxplot.}

\item{log}{whether log2 transform the exp data.}

\item{order}{default=TRUE, fix the order of sample in the boxplot.}
}
\value{

}
\description{
Title
}
