% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_normalization_variable-specifc.R
\name{normalize_metabolite_specific}
\alias{normalize_metabolite_specific}
\title{variable-specific normalization (QC-based signal correction)}
\usage{
normalize_metabolite_specific(
  expData,
  design,
  method,
  rf.ntree = 500,
  loess.span = 0.5,
  loess.degree = 2,
  ...
)
}
\arguments{
\item{expData}{data.frame: row as sample, colum as metabolite(or features/variables)}

\item{design}{data.frame, including 4 columns: "sample", "batch", "class", "order" (run order). QC sample as "QC" in column "class".}

\item{method}{QC-based signal correction methods, including: qc.mean, qc.median, loess (QCRLSC), random forest (QCRFSC), QC-SVR, IORLSC, IORFSC}

\item{rf.ntree}{ntree par for randomForest::randomForest()}

\item{loess.span}{span par for loess()}

\item{loess.degree}{degree par for loess()}

\item{...}{}
}
\value{

}
\description{
variable-specific normalization (QC-based signal correction)
}
\examples{

}
