% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_normalization_sample-based.R
\name{normalize_sample_based}
\alias{normalize_sample_based}
\title{Sample-based normalization
Normalise data by sample-based scaling factor, like sum or mean of sample.}
\usage{
normalize_sample_based(expData, design, method, na = 0)
}
\arguments{
\item{expData}{data.frame: row as sample, colum as metabolite(or features/variables)}

\item{design}{data.frame, including 4 columns: "sample", "batch", "class", "order" (run order). QC sample as "QC" in column "class".}

\item{method}{including: "tic", "median", "mean", "range", "mad", "MSTUS".}

\item{na}{only take 0 as missing value, not NA.}
}
\value{

}
\description{
Sample-based normalization
Normalise data by sample-based scaling factor, like sum or mean of sample.
}
