% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_consensus_clusters.R
\name{plot_consensus_cluster}
\alias{plot_consensus_cluster}
\alias{plot_consensus_cluster.heatmap}
\alias{plot_consensus_cluster.pca}
\alias{plot_cluster_algorithms.pca}
\alias{plot_cluster.pca}
\title{plot_consensus_cluster}
\usage{
plot_consensus_cluster.heatmap(
  res.con,
  clustering_method_columns = "ward.D2",
  clustering_method_rows = NULL
)

plot_consensus_cluster.pca(res.con)

plot_cluster_algorithms.pca(res.con)

plot_cluster.pca(res.con, method)
}
\arguments{
\item{res.con}{Result of Consensus_Cluster_Analysis().}

\item{clustering_method_columns}{Hierarchical clustering used for clustering differnt clusters result of multi-clustering algorithm.}

\item{clustering_method_rows}{Default as same as consensus.method used in \code{Consensus_Cluster_Analysis()}. Can be set as any mehthods in \code{hclust}}

\item{method}{clustering method list in \code{clustering.methods()}.}
}
\description{
Visualiztion of consensus result
}
\details{
plot_consensus_cluster
}
\examples{
plot_consensus_cluster.heatmap(res.con)
plot_consensus_cluster.pca(res.con)

plot_cluster_algorithms.pca(res.con)
plot_cluster.pca(res.con, "kmeans")
}
