% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_consensus_clusters.R
\name{plot_consensus_clusters}
\alias{plot_consensus_clusters}
\alias{plot_consensus_clusters.heatmap}
\alias{plot_consensus_clusters.pca}
\title{plot consensus result}
\usage{
plot_consensus_clusters.heatmap(consensus_clusters)
}
\arguments{
\item{consensus_clusters}{Result of Consensus_Cluster_Analysis().}
}
\description{
plot consensus result
}
