% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_cellCycleScoring.R
\name{SC_cellCycleScoring}
\alias{SC_cellCycleScoring}
\title{Score cell cycle phases
Cell cycle classification including G2/M, S, G1.}
\usage{
SC_cellCycleScoring(
  counts,
  method = "Seurat",
  graph.pca = TRUE,
  cc.genes = NULL,
  normalization.method = "LogNormalize"
)
}
\arguments{
\item{counts}{raw counts matrix: row as genes, column as cells/samples. rownames must be gene SYMBOL.}

\item{method}{one of: "Seurat",}

\item{graph.pca}{whether to show pca graph which only based on 97 cell cycle genes or not.}

\item{cc.genes}{a list, with cc.genes and s.genes vectors. Defult are cc.genes from Seurat package.}

\item{normalization.method}{Method for counts normalization: "LogNormalize", "CLR", "RC". see \code{\link[Seurat]{NormalizeData}}}
}
\value{
factor vector for classfication (G1, G2/M, S)
}
\description{
Score cell cycle phases
Cell cycle classification including G2/M, S, G1.
}
\seealso{
\link{https://satijalab.org/seurat/v3.2/cell_cycle_vignette.html}
}
