% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_clusters.R
\name{cluster_number}
\alias{cluster_number}
\title{cluster_number}
\usage{
cluster_number(
  Exp,
  k.max,
  FUNcluster = c(kmeans, pam, clara, fanny, hcut),
  scale = TRUE,
  ...
)
}
\arguments{
\item{Exp}{data.frame with rownames represent samples, colnames represent features like proteins or genes.}

\item{k.max}{maximum numbers of cluster.}

\item{FUNcluster}{Clustering method. Type without quote mark. Chose one of following methods: kmean, pam, clara, fanny, hcut.}

\item{scale}{Logical. whether input Exp should be scaled.}

\item{...}{Any parameter passed to fviz_nbclust().}
}
\value{
Two plots. Indicating number of clusters by using Elbow method and Silhouette method.
}
\description{
Produce one referenced number of clusters which determined by kmean/pam/clara/fanny/hcut clustering.
}
\examples{
cluster_number(iris[-5], 5, kmeans)

}
\seealso{
\code{\link[cluster]{fviz_nbclust}}
}
