% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA_plot_consensus_clusters.R
\name{plot_consensus_cluster}
\alias{plot_consensus_cluster}
\alias{plot_consensus_cluster.heatmap}
\alias{plot_consensus_cluster.pca}
\alias{plot_cluster_algorithms.pca}
\alias{plot_cluster.pca}
\alias{plot_exp_heatmap}
\title{plot_consensus_cluster}
\usage{
plot_consensus_cluster.heatmap(
  res.con,
  main = "clutering algorithms",
  clustering_method_columns = "ward.D2",
  clustering_method_rows = NULL
)

plot_consensus_cluster.pca(res.con, main = "Consensus cluster")

plot_cluster_algorithms.pca(res.con)

plot_cluster.pca(res.con, method)

plot_exp_heatmap(
  res.con,
  method = "ALL",
  clustering_method_columns = "complete",
  clustering_method_rows = "complete",
  col_title_prefix = "Subclass",
  top_annotation = NULL,
  right_annotation = NULL
)
}
\arguments{
\item{res.con}{Result of Consensus_Cluster_Analysis().}

\item{main}{Title of plot.}

\item{clustering_method_columns}{Hierarchical clustering method used for clustering differnt clusters result of multi-clustering algorithm, pass to \code{\link[stats]{hclust}}.}

\item{clustering_method_rows}{Default as same as consensus.method used in \code{Consensus_Cluster_Analysis()}, pass to \code{\link[stats]{hclust}}.}

\item{method}{clustering method list in \code{clustering.methods()}.}

\item{col_title_prefix}{Prefix for colum subclass name.}

\item{top_annotation}{A HeatmapAnnotation-class generated by \code{\link[ComplexHeatmap]{HeatmapAnnotation}}, annotation information of column(here is sample).}

\item{right_annotation}{A HeatmapAnnotation-class generated by \code{\link[ComplexHeatmap]{HeatmapAnnotation}}, annotation information of row(here is variable).}
}
\description{
Visualization of consensus result
}
\details{
plot_consensus_cluster
}
\examples{
plot_consensus_cluster.heatmap(res.con)
plot_consensus_cluster.pca(res.con)

plot_cluster_algorithms.pca(res.con)
plot_cluster.pca(res.con, "kmeans")

#Use consensus cluster result for column subclass on heatmap.
plot_exp_heatmap(res.con)
#Use specific method result like kmeans for column subclass on heatmap
plot_exp_heatmap(res.con, "kmeans")
}
